+++ src/either/model.ts
@@ -6,4 +6,8 @@ export type Right<T> = {
 	readonly right: T;
 };
 
-export type EitherV<E, T> = Left<E> | Right<T>;
+export type EitherValue<E, T> = Left<E> | Right<T>;
+
+export type EitherResult<E, T> =
+	| { ok: true; value: T }
+	| { ok: false; error: E };