+++ src/validation.ts
@@ -1,29 +1,4 @@
-import type { Either } from "./either.js";
-
-export type ValidationV<_E, _T> = object;
-
-export type Validation<_E, _T> = {
-	ap: unknown;
-	map: unknown;
-	map3: unknown;
-};
-
-export function valid<T, E = never>(): Validation<E, T> {
-	throw new Error("Not implemented");
-}
-
-export function invalid<E, T = never>(): Validation<E, T> {
-	throw new Error("Not implemented");
-}
-
-export function fromEither<E, T>(_either: Either<E, T>): Validation<E, T> {
-	throw new Error("Not implemented");
-}
-
-export function traverse() {
-	throw new Error("Not implemented");
-}
-
-export function sequence() {
-	throw new Error("Not implemented");
-}
+export * from "./validation/constructors.js";
+export * from "./validation/helpers.js";
+export * from "./validation/model.js";
+export type { Validation } from "./validation/validation.js";